intl/l10n/test/dom/test_mozdom_translateFragment.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test mozIDOMLocalization.translateFragment</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  const { getDOMLocalization } =
    ChromeUtils.import("resource://gre/modules/DOMLocalization.jsm");
  const domLocalization = getDOMLocalization();

  const { Services } = ChromeUtils.import("resource://gre/modules/Services.jsm");
  const { L10nRegistry, FileSource } =
    ChromeUtils.import("resource://gre/modules/L10nRegistry.jsm");

  const fs = {
    "/localization/en-US/browser/menu.ftl": `
title = Hello World
subtitle = Welcome to Fluent
`,
  };
  const originalLoad = L10nRegistry.load;
  const originalRequested = Services.locale.requestedLocales;

  L10nRegistry.load = async function(url) {
    return fs.hasOwnProperty(url) ? fs[url] : false;
  };

  const source = new FileSource("test", ["en-US"], "/localization/{locale}");
  L10nRegistry.registerSource(source);

  window.onload = async function() {
    SimpleTest.waitForExplicitFinish();

    domLocalization.addResourceIds(["/browser/menu.ftl"], false);

    const frag = document.querySelectorAll("div")[0];
    const h1 = document.querySelectorAll("h1")[0];
    const p1 = document.querySelectorAll("p")[0];

    await domLocalization.translateFragment(frag);
    is(h1.textContent, "Hello World");
    is(p1.textContent, "Welcome to Fluent");

    // Cleanup
    L10nRegistry.removeSource(source.name);
    L10nRegistry.load = originalLoad;
    Services.locale.requestedLocales = originalRequested;

    SimpleTest.finish();
  };
  </script>
</head>
<body>
  <div>
    <h1 data-l10n-id="title" />
    <p data-l10n-id="subtitle" />
  </div>
</body>
</html>
